Track move functionality, code cleanup
Changes include:
-Added more scaffolding for moving tracks. Tracks can now be moved up
and down in the order, and their new positions are saved with the
project. Undo/redo history is also saved for move operations.
-Moving tracks is currently very buggy and can only be done once before
the application becomes unstable. A transform animation is used to move
the tracks onscreen, but their order in the DOM is not currently
updated. Click handlers for the "move" buttons tend to duplicate
themselves each time the "move" flyout is shown.
-Assorted cleanup of some CSS classes/IDs (focusing on the BEM
methodology going forward).
-Added credits.txt, containing credits to be included in the final
version of the application.