Numerous changes in preparation for media playback
Change include:
-Added PlaybackMGR scaffolding.
-Added Clip class scaffolding.
-Cleaned up some unused classes in default.css and editor.css.
-Removed old, irelevant comments from default.html.
-Refactored Ensemble.Editor.Pages.Editor.UI to Ensemble.Editor.UI.
-Split a number of the click handlers out of editor.js and into their
respective components (TimelineMGR has its own ui object now and keeps
track of its own references). Slightly altered the project load/unload
chain to be more compartmentalized and generally less cluttered across
the board.
-Added Clip loading from the MediaBrowser's preview window. Opening a
preview and selecting "Add to project" presents the user with a list of
all available Tracks in the project. Selecting a track triggers the
loading of the media file in Ensemble.FileIO. A system of callbacks will
be used to deliver the loaded Clip object back to the MediaBrowser
(callback currently does not fire; this is intentional - the loading
system isn't ready for it yet).
-Disabled some of the Editor's animations (track vertical scrolling,
sorting tracks) because they weren't working very well. Proper
animations will be added at a later date.
-Removed "editorTimelineRowsVisible" from the application settings. This
user setting has been outmoded by the TimelineMGR, which now
intelligently sizes tracks to best fit the Timeline display.
-Added a subtle background texture to the loading screens and updated
credits.txt to include a reference to the source of the new graphic.