More media playback preparation
Changes include:
-Finished the Clip backend loading system for importing media from the
MediaBrowser's preview dialog. The loaded Clip object is delivered all
the way back to the TimelineMGR in a state ready for playback: now the
TimelineMGR just needs to build an appropriate rendering target for the
PlaybackMGR and the Renderer.
-Generated a new package ID because VS was complaining about it. This
change is trivial.
-Ensemble of One project files (.eo1) now maintain the next free unique
ID for Clip objects. This is in preparation for media importing.