Added timing mechanism, playback, rendering
Changes include:
-Added a basic ability to play back a project.
-Created a new timing system that runs in a Worker (timing updates are
processed on a separate thread from the UI thread). The timing system
fires events back to the PlaybackMGR, which uses them to start/stop the
appropriate clips as well as determine when playback has reached the end
of a project.
-Added Renderer. A single frame is rendered after a clip is imported.
The Renderer can also be started and stopped; when in the "active"
state, the Renderer will continuously schedule new frame redraws. A
single frame can be requested at any time by calling
Ensemble.Editor.Renderer.renderSingleFrame().
-Tweaked the Media Browser so that its dialogs close automatically when
a clip is imported into the project.
-Fixed a naming error in Editor.js that was probably created when
Ensemble.Pages.Editor.UI was refactored to Ensemble.Editor.UI.