Project load test code, FileIO refactoring, seek
Changes include:
-Added a seek(value) method to the PlaybackMGR namespace. Pass it any
value in milliseconds, and the project's current playback position will
be seeked to that point. Time values outside the project's duration are
clamped.
-General bugfixes to the timing mechanism.
-Wrote backend test code for the project load operation. Attempting to
load a project will now trigger the load of all the clips saved inside
it. As a result, the "loading" screen will never dismiss for projects
that contain clips. This will be fixed in the next update.
-General cleanup of the FileIO namespace. Added a fix for dynamic blob
URI generation that will hopefully make loading clips a little more
stable (clips were sometimes silently "dropped" if the app already had
references to them in a previous run).