Connected FileIO project load chain to Editor
Changes include:
-Projects are now properly loaded. The Editor screen automatically shows
when loading completes. Although Track objects are being fully recreated
upon project load, they are not currently displayed in the Timeline.
This will be fixed in a future update.
-Moved Editor init calls to before any entrance animations are played.
That way, we can count on already having a frame on the canvas and that
the Timeline is properly built before they enter the screen.
-Removed an old relic from the "set number of timeline rows visible"
setting.
-Project loading is currently done via absolute file paths. No error
handling occurs: if a media file fails to load, Ensemble of One will get
"stuck" on the loading screen and wait forever. This will be corrected
in a future update.