Moved clip import code into Action
Changes include:
-Moved clip import code into the Action class, instead of its original
location in the Media Browser. Now an Action can be used to
independently trigger the load of a media clip from a Clip stub.
-Known issue: imported clips do not set their duration upon initial load
because retrieveMediaProperties() is not called. Media property
retrieval will be chained into Clip importing in the next update.
-Added a helper function (finish()) to the Action class to help with
cleanup after an Action needs to trigger a chain of asynchronous
operations (such as file IO).
-History is now properly saved for clip imports. Clips cannot currently
be removed, and undoing a clip import Action may result in application
instability.