Added seek UI, recent history message
Changes include:
-Added draggable cursor to the timeline. The project is seeked when the
user drags the curor and lets go of it. Cursor automatically moves
during playback and may move out of view if the user does not scroll.
-Added a persistent history message that displays the last
user-committed Action (this is the Action that will be undone if the
user clicks "Undo" in the menu or presses the "Ctrl+Z" key combination).
-Fixed a bug that caused the Renderer to crash upon attempting to render
a frame with no clips in it (base case: new project with no clips).
-Fixed a syntax error in Ensemble.Utilities.TimeConverter that prevented
IntelliSense from displaying proper autocomplete information in Visual
Studio.