Added browse dialog to Editor
Changes include:
- Closed #41. User can now use the "Browse" button on the Editor toolbar
to manually browse for and import a clip. After selecting a clip from
the standard OS dialog, a dialog is presented allowing the user to
select a destination track. Dismissing the dialog cancels the import
without affecting the application cache.
- Added FutureAccessList management to file browsing. When the user
tries to import a file that would exceed the capacity of the
FutureAccessList, Ensemble of One will automatically "expire" the oldest
item off the list.