UI, media playback fixes for Windows 10
Changes include:
- Additional fixes for media playback in Windows 10. Clips are now
physically appended to the DOM on project load, and subsequently removed
from the DOM if the clip is removed from the project or the project is
unloaded.
- Cleanup and simplification of Editor code, including reorganizing
HTML/CSS revisions to contend with the Edge engine's weird flexbox
quirks.
- Removed user-resizable horizontal break from the Editor because it was
having so much trouble with the new model. It may be implemented again
at a later date.