Enabled sticky edges for resize operations
Changes include:
- Enabled the "sticky edges" feature for clip resize operations. Resized
clips will now stick to other clips.
- Resizing sticky edges has several known bugs, the most outstanding
being that the resized clip tends not to stick when its dimensions
exceed those of the nearby clip to which the user is trying to stick it.
- There's some code duplication with the sticky edges for "move"
operations; this will be fixed once the resize-snap code is debugged and
cleaned up a little bit.
- Various refactoring and reorganization of Ensemble.Editor.Renderer for
clip resizing, moving, and snapping.
- Changed the color scheme for the Windows title bar controls so it's a
little easier to tell when a control is being hovered or clicked.
- Main Menu now only enumerates projects when showing the "Open project"
page (it previously enumerated projects on every menu page switch, which
was both overkill and redundant).
- Revised the order of operations and animation sequence for the "Open
project" Main Menu page. Projects now appear more quickly and more
smoothly than previously.