Audio support, Timeline enhancements, bug fixes
Changes include:
- Closed #70. Users can now click any location in the Timeline ruler to
instantly seek to that position.
- Closed #55. Audio files can now be imported into user projects, and do
not render to the Canvas.
- Various updates to Ensemble.Editor.Renderer,
Ensemble.Editor.SelectionMGR, and Ensemble.Editor.TimelineMGR to
properly support audio files.
- Updates to Ensemble.Editor.SelectionMGR to increase efficiency and
decrease the likelihood that a dead frame will be requested (a frame in
which nothing has changed).
- Closed #20. Added a new case to timer.js to handle "seek to zero"
situations.
- Closed #25. Mostly just considering it closed, since it's been
effectively closed and revised in previous commits. Take note of
outstanding bug #60.
- Updated Ensemble.Editor.TimelineMGR to separate renderable clips from
non-renderable clips in the playback index. This helps prevent
unnecessary rendering passes and was necessary for adding audio file
support.