Massive code cleanup and consistency changes
Changes include:
- Removed Ensemble.KeyboardMGR.
- Removed Ensemble.Pages.MainMenu (replaced with Ensemble.MainMenu).
- Closed #80.
- Removed [currently] unused properties from the Ensemble.Editor.UI
object (some items still remain and are in use, but will be refactored
out in a later update).
- Closed #78. Editor title bar commands have now been changed into HTML
<button> elements, and Ensemble.Editor.MenuMGR has been updated to
reflect this change.
- Began work on #81. The old DOM-based ruler has been temporarily
disabled in favor of the optimized, <canvas>-based upcoming replacement.
- Began work on #74. Disabling the old Timeline ruler solved most of the
problem, but there are still opportunities for improvement.
- Began work on #64. All of the old clickeaters have been removed, with
the exception of "editorTimelineEWClickEater." The issue will be closed
once this final clickeater has been removed.
- Further investigation required on #61, which now appears to be an
issue with the Edge HTML/JS engine.
- Temporarily disabled loading a project by double-clicking its file in
Windows Explorer. This feature was broken when the new Main Menu was
implemented, and will be replaced via #56 anyway.
- Removed obsolete code from the Media Browser.
- Removed obsolete code from default.js that pertained to the extended
splash screen from Windows 8.1.
- Various fixes from the WinJS 4.0.0 release update that had previously
been missed.
- Minor changes to the Timeline button styles in the Editor, pending
further refinement.
- Removed basically everything from default.css. Only a few essential
items now remain.
- Minor edits to the Settings menu animation, so it properly obstructs
the Editor view when activated.