Cleanup and scaffolding for Editor widgets
Changes include:
- Removed a couple entries from Ensemble.Editor.UI and moved them to
Ensemble.Pages.Editor to be more compartmentalized.
- Corrected some of the overflow positioning of the playback canvas and
its controls (the upper half contents previously overflowed their bounds
by a small amount due to an incorrect calculation).
- Scaffold code preparation for #30, #33, and #86. The playback canvas
and its controls are now part of the same container, making it easier to
displace it in the event a mini-app is open.