Remember last webcam quality, match camera preview
Changes include:
- Continued working on #30.
- Ensemble of One now remembers the last used camera quality and
attempts to load it when the user initializes a camera. If the quality
cannot be matched (due to a hardware change, etc.), then webcam capture
defaults to the highest quality available on the capture device.
- Camera preview now matches the capture quality and aspect ratio more
intelligently. Ensemble of One now sets the webcam preview to the stream
quality that most closely matches the aspect ratio of the capture
stream, with as high a resolution as possible (without exceeding that of
the capture stream, that is).