Added HistoryMGR batch operations
Changes include:
- Continued working on #30.
- Added a slight delay in between starting preview and starting capture,
since Surface still seems to hang on occasion when starting a new
capture session.
- Updated the HistoryMGR to support batch operations. You can now pass
an array of Action objects and have them all executed sequentially,
firing a callback when all have completed.
- Batch Actions are a little buggy: the callback doesn't seem to fire at
the end, and the project is not saved after all Actions in the batch
have completed. These issues will be fixed in a future update.