Webcam capture stability fixes, various UI tweaks
Changes include:
- Closed #110.
- Changed all red Eo1 buttons to have rounded corners and removed
customizable corners due to the buggy CSS implementation.
- Minor coloration changes to the Editor, including a new background
color for the Timeline and for the video capture popin's "import" list
overlay.
- Updates to the webcam capture "import" list items. List items now also
display the time at which they will import, as well as their estimated
capture duration. Configuration options (such as where to import, as
well as the option to tweak the import time) will be included in a
future update, likely in a pop-out Flyout overlay.
- Adjusted Timeline container padding so the main control buttons are
not right up against the edge of the application window.
- Removed WebGL test code from Ensemble.Editor.Clip (existing code was
generating errors and didn't actually work correctly, anyway).