Timeline layout fixes and scroll capabilities
Changes include:
- Closed #119. There is still some cleanup to do in the refactoring
department, as well as some missing features on the Timeline. Drag &
drop is probably broken, and will likely cause a fatal error if
attempted.
- Added vertical scrolling to Timeline via mouse wheel. See bug #120.
- Updated time cursor so it is always visible (vertically), regardless
of the user's vertical scroll position.
- Vertical scrolling plays a short animation, and also snaps to whole
Tracks. The sizing of the Timeline itself will be adjusted in a future
update so the little "sliver" of track visible at the bottom on some
screen sizes is no longer an issue.