Project ListView and loading improvements
Changes include:
- Continued work on #100. "Recent" and "unsaved" projects now display
together in a single ListView, separated by category headers.
- Removed the "New" and "Open" menus on the main menu. All project
creation and loading can now be done from the main screen.
- Made the main menu headings a little smaller so they're not quite so
enormous.
- Cleaned up redundant project-loading code in Ensemble.FileIO. Projects
are now always loaded via a StorageFile (previously, "unsaved" projects
were loaded via filename instead and assumed to be stored in the AppData
directory).
- Added property to Ensemble.ProjectFile indicating whether the project
is "unsaved" or is external to the application.