MediaComposition prep and XML cleanup
Changes include:
- Basic preparations for implementing the Windows 10 MediaComposition
API (filler images, Ensemble.FileIO organization).
- Converted project creation code to use native JS implementation to
generate an XML document (rather than the 3rd-party library that was
previously used). This change has not yet been made to other save/load
sections of the app, so project loading is broken for the time being.
- Corrected "2160p" to read "UHD 4K" (which is a more useful description
and is the common convention for that resolution).